# storage/api-docs/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Ремонт часов на Алексеевской",
        "contact": {
            "name": "Анастасия Паршенкова",
            "email": "parshenkova.anastasya@yandex.ru"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/statuses/": {
            "get": {
                "tags": [
                    "Status"
                ],
                "description": "Возвращает перечень статусов",
                "operationId": "b0ce668e879db23f3aca6cd047272eca",
                "responses": {
                    "200": {
                        "description": "Перечень статусов",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Статус не найден",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Status"
                ],
                "description": "Создаёт статус",
                "operationId": "4e6de9f71f0d66e49524b2b9afb1d60b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#/components/schemas/Type"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Статус создан",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/statuses/{status}/": {
            "get": {
                "tags": [
                    "Status"
                ],
                "description": "Возвращает статус по идентификатору",
                "operationId": "c7624059bef7d84d3c346cef01c73249",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/status"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Статус",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Статус не найден",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Status"
                ],
                "description": "Удаляет запись о статусе",
                "operationId": "3c3bb30d470c1108299146c8b6e1aa31",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/status"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Запись о статусе удалена",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "Status"
                ],
                "description": "Корректирует запись о статусе",
                "operationId": "76d6eb1652d1d8eab7f80eb626dde6ca",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/status"
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#/components/schemas/Status"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "Запись о стастусе откорректирована",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/types/": {
            "get": {
                "tags": [
                    "Type"
                ],
                "description": "Возвращает перечень типов",
                "operationId": "b494032c6d3227acb4b5ff2f6684411b",
                "responses": {
                    "200": {
                        "description": "Перечень типов",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Тип не найден",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Type"
                ],
                "description": "Создаёт тип",
                "operationId": "99078134d285601ed8fd1f4ba4a04361",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#/components/schemas/Type"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Тип создан",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/types/{type}/": {
            "get": {
                "tags": [
                    "Type"
                ],
                "description": "Возвращает тип по идентификатору",
                "operationId": "5f1f4f6e35493e875a4b35a0fd9d2c90",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/type"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Тип",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Тип не найден",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Type"
                ],
                "description": "Удаляет запись о типе",
                "operationId": "23f0d04f7d86d132409da86599741067",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/type"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Запись о типе удалена",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "Type"
                ],
                "description": "Корректирует запись о типе",
                "operationId": "5013003212006b65de27e66062b138ca",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/type"
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#/components/schemas/Type"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "Запись о типе откорректирована",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Status": {
                "title": "Status",
                "required": [
                    "name"
                ],
                "properties": {
                    "name": {
                        "description": "Название статуса",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "Type": {
                "title": "Type",
                "required": [
                    "name"
                ],
                "properties": {
                    "name": {
                        "description": "Название типа",
                        "type": "string"
                    }
                },
                "type": "object"
            }
        },
        "parameters": {
...
            "status": {
                "name": "status",
                "in": "path",
                "description": "Первичный ключ"
            },
            "type": {
                "name": "type",
                "in": "path",
                "description": "Первичный ключ"
            }
        }
    },
...
}